requirements.txt
requests
black
ffmpeg
ffsubsync
thefuzz
pyyaml
rich